isolate search app js entrypoint
