+++ oarepo_vocabularies/ui/theme/webpack.py
@@ -7,7 +7,7 @@ theme = WebpackThemeBundle(
     themes={
         "semantic-ui": dict(
             entry={
-                "oarepo_vocabularies_search": "./js/oarepo_vocabularies_ui/search/index.js",
+                "oarepo_vocabularies_search": "./js/oarepo_vocabularies_ui/search/app.js",
                 "oarepo_vocabularies_ui_components": "./js/oarepo_vocabularies_ui/custom-components.js",
                 "oarepo_vocabularies_detail": "./js/oarepo_vocabularies_ui/detail/index.js",
                 "oarepo_vocabularies_form": "./js/oarepo_vocabularies_ui/form/app.js",
